arrays
The task of the work: to write an application that allows the user to enter N numbers of an array into the console (the number N is also entered by the user). The program then outputs all the values ​​into the console and then outputs the minimum and maximum numbers.